requirements.txt
bleach==2.1.3
cycler==0.10.0
Cython==0.29.13
decorator==4.2.1
entrypoints==0.2.3
future==0.17.1
hdrhistogram==0.6.1
html5lib==1.0.1
ipykernel==4.8.2
ipyparallel==6.1.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2>=2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client>=5.3.1
jupyter-console==5.2.0
jupyter-core>=4.6.1
kiwisolver==1.0.1
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook>=5.7.2
numpy>=1.18.1
pandas>=0.23.3
pandocfilters==1.4.2
parso==0.1.1
pbr==5.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
tqdm==4.24.0
traitlets==4.3.2
virtualenv==16.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
